add examples + test adaptation to new functionalities + outsorcing utils for yaml search space in own file
